[SYSTEMDS-3487] Array primitives with null
This commit adds a new Primitive super class for the arrays, to enable
the primitive arrays such as int[] to contain nulls.
This in essence allows us to allocate columns containing null values
in the correct datatype without having being able to use null in the
individual values.

In the process of adding this new primitive many things broke, since
they do not support having null values inside int or double arrays.
To fix these some tests were modified, and remove empty was moved from
FrameBlock to a lib: FrameBlockLibRemoveEmpty.

The readers that did not support null now does: Json Reader, and Proto
write and read, and text format now writes out sparsely to not write out
default values on columns. For instance zero values in an int array column.

Closes #1763

+++ src/test/java/org/apache/sysds/test/component/compress/io/IOCompressionTestUtils.java
@@ -34,9 +34,12 @@ public class IOCompressionTestUtils {
 
 	static final AtomicInteger id = new AtomicInteger(0);
 
-	protected static void deleteDirectory(File file) {
+	public static void deleteDirectory(File file) {
 		synchronized(IOCompressionTestUtils.lock) {
-			for(File subfile : file.listFiles()) {
+			File[] files = file.listFiles();
+			if(files == null)
+				return;
+			for(File subfile :files) {
 				if(subfile.isDirectory())
 					deleteDirectory(subfile);
 				subfile.delete();
